Refactor all components on Hotel Details Page
