+++ src/components/ReviewsCarousel.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { Typography, Grid } from "@mui/material";
 import Slider from "react-slick";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
@@ -13,10 +14,17 @@ export default function ReviewsCarousel({ reviews }) {
     slidesToScroll: 3,
   };
   return (
-    <Slider {...settings}>
-      {reviews.map((review) => (
-        <Review key={review.id} review={review} />
-      ))}
-    </Slider>
+    <Grid container mb={7}>
+      <Typography variant="h6" component="h2" fontWeight="600" pb={2}>
+        Reviews
+      </Typography>
+      <Grid item xs={12} md={12} lg={12}>
+        <Slider {...settings}>
+          {reviews.map((review) => (
+            <Review key={review.id} review={review} />
+          ))}
+        </Slider>
+      </Grid>
+    </Grid>
   );
 }
